refactor(ui): regroup typography under primitives/static
