+++ packages/ui-react/src/components/primitives/static/Typography/index.ts
@@ -1,0 +1,2 @@
+export type { TypographyProps } from './Typography';
+export { Typography } from './Typography';
