+++ packages/ui-vue/src/components/primitives/static/Typography/index.ts
@@ -1,0 +1,1 @@
+export { default, default as Typography } from './Typography.vue';
