+++ packages/ui-vue/src/index.ts
@@ -13,4 +13,4 @@ export {
   tokenThemes,
   typographyVariants,
 } from '@hoite-dev/ui';
-export { default as Typography } from './components/Typography.vue';
+export { default as Typography } from './components/primitives/static/Typography';
